requirements-dev.txt
# ===================================================
# DataFlow Pro - Development Dependencies
# ===================================================
# Install: pip install -r requirements-dev.txt
# ===================================================

# Include production dependencies
-r requirements.txt

# Testing
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
pytest-mock>=3.12.0,<4.0.0

# Linting & Formatting
ruff>=0.1.0,<1.0.0
black>=23.0.0,<25.0.0

# Type Checking
mypy>=1.7.0,<2.0.0
pandas-stubs>=2.0.0

# Security Auditing
pip-audit>=2.6.0,<3.0.0
